channels/pyproject.toml
[project]
name = "channels"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "supabase>=2.11.0",
]
